requirements.txt
streamlit==1.32.2
tensorflow==2.10.1
keras==2.10.0
protobuf==3.20.3
Pillow
gtts
transformers
